+++ gulpfile.js
@@ -2,7 +2,6 @@
 const gulp = require('gulp');
 const bump = require('gulp-bump');
 const package = require('./package.json')
-var gulp = require('gulp');
 
 gulp.task('copyLocaleFromLib', done => {
   gulp
@@ -16,21 +15,9 @@ gulp.task('watch:locale', function () {
   gulp.watch('./node_modules/@igo2/core/locale/*.json', gulp.series('copyLocaleFromLib'));
 });
 
-gulp.task('watch:locale', function () {
-  gulp.watch('./node_modules/@igo2/core/locale/*.json', gulp.series('copyLocaleFromLib'));
-});
-
 gulp.task('bumpPwaVersionDev', done => {
   gulp.src(['./ngsw-config.json'])
     .pipe(bump({ type: 'prerelease' }))
-    .pipe(gulp.dest('./'));
-  done();
-});
-
-gulp.task('bumpPwaVersionProd', done => {
-  const packageVersion = package.version;
-  gulp.src(['./ngsw-config.json'])
-    .pipe(bump({ version: packageVersion+'-pwa' }))
     .pipe(gulp.dest('./'));
   done();
 });